fixed no stream file writing

+++ api/src/inference/base.py
@@ -21,7 +21,7 @@ class AudioChunk:
         output=AudioChunk(audio_chunk_list[0].audio,audio_chunk_list[0].word_timestamps)
         
         for audio_chunk in audio_chunk_list[1:]:
-            output.audio=np.concatenate((output.audio,audio_chunk.audio))
+            output.audio=np.concatenate((output.audio,audio_chunk.audio),dtype=np.int16)
             if output.word_timestamps is not None:
                 output.word_timestamps+=output.word_timestamps
